+++ custom_components/unraid_api/const.py
@@ -7,7 +7,7 @@ from typing import Final
 from homeassistant.const import Platform
 
 DOMAIN: Final = "unraid_api"
-PLATFORMS = [Platform.SENSOR, Platform.BINARY_SENSOR]
+PLATFORMS = [Platform.SENSOR, Platform.BINARY_SENSOR, Platform.SWITCH]
 
 CONF_SHARES: Final[str] = "shares"
 CONF_DRIVES: Final[str] = "drives"